Add benchmarks for Simulator

+++ simulator/src/simple/movable.rs
@@ -9,7 +9,7 @@ pub struct RandPerson {
 impl Movable for RandPerson {
     fn get_speed(&self) -> f32 {self.speed}
     fn set_speed(&mut self, s: f32) {self.speed = s}
-    fn update(&mut self, t: f32) {}
+    fn update(&mut self, t: f64) {}
     fn decide_next<'a>(&mut self, connections: &'a Vec<usize>) -> &'a usize {
         let i = rand::thread_rng().gen_range(0..connections.len());
         &connections[i]
@@ -32,7 +32,7 @@ impl RandCar {
 impl Movable for RandCar {
     fn get_speed(&self) -> f32 {self.speed}
     fn set_speed(&mut self, s: f32) {self.speed = s}
-    fn update(&mut self, t: f32) {}
+    fn update(&mut self, t: f64) {}
     fn decide_next<'a>(&mut self, connections: &'a Vec<usize>) -> &'a usize {
         let i = rand::thread_rng().gen_range(0..connections.len());
         &connections[i]
